refactor(mcp/write): #464 extract CLUSTER_B AST metrics into write_metrics.py (STRATEGY_S1)
Mechanical VERBATIM extraction (step 2 of 4) of the STRATEGY_S1 sequence
under parent issue #459. Predecessor: PR #467 (CLUSTER_A — write_detection.py).

Symbols moved from src/octave_mcp/mcp/write.py to new module
src/octave_mcp/mcp/write_metrics.py:
- StructuralMetrics dataclass
- extract_structural_metrics function (including nested traverse helper)

write.py now imports both symbols from write_metrics. No re-export shim
(parent #459 acceptance). Test file imports updated to the new module path.
Unused `from dataclasses import dataclass, field` removed from write.py.

LOC delta:
- src/octave_mcp/mcp/write.py: 3961 → 3917 (-44)
- src/octave_mcp/mcp/write_metrics.py: new, 55 LOC (40 LOC of code + docstring + imports)

Quality gates:
- ruff: clean
- black: clean
- mypy: clean
- pytest: 3614 passed, 11 skipped, 3 xfailed (matches post-CLUSTER_A baseline, zero regressions)

PROD::I1 SYNTACTIC_FIDELITY preserved — no signature change, no body edit,
no rename. Pure structural relocation.

Refs: #459 #464
Predecessor: #463 / PR #467
